+++ src/TSDB_store.c
@@ -335,7 +335,7 @@ int TSDB_last_row_idx (struct TSDB_metric_t *obj, int column) {
 }
 
 
-int TSDB_cell (struct TSDB_metric_t *obj, int row, int column) {
+u_int64_t TSDB_cell (struct TSDB_metric_t *obj, int row, int column) {
   char *buf;
   u_int64_t rv;
   int read_len, offset;
